+++ CHANGES.md
@@ -27,6 +27,8 @@
   ([#126](https://github.com/semgrep/testo/pull/126)).
 * Add support for timeouts
   ([#127](https://github.com/semgrep/testo/issues/127)).
+* Add support for checked output files
+  ([#134](https://github.com/semgrep/testo/issues/134)).
 
 0.1.0 (2024-11-10)
 ------------------